fix test again

+++ common/persistence/elasticsearch/esVisibilityStore_test.go
@@ -137,7 +137,7 @@ func (s *ESVisibilitySuite) TestRecordWorkflowExecutionStarted() {
 		s.Equal(request.IsCron, fields[es.IsCron].GetBoolData())
 		s.Equal((int64)(request.NumClusters), fields[es.NumClusters].GetIntData())
 		s.Equal(indexer.VisibilityOperationRecordStarted, *input.VisibilityOperation)
-		s.Equal(request.ShardID, fields[es.ShardID].GetIntData())
+		s.Equal((int64)(request.ShardID), fields[es.ShardID].GetIntData())
 		return true
 	})).Return(nil).Once()
